Schedulers of particle effects are now handled by the superclass
